// types/theme/theme.ts
export interface colorThemeTypes {
  primaryFontColor: string;
  secondaryFontColor: string;
  pendingColor: string;
  paidColor: string;
  formTitleColor: string;
  formLabelColor: string;
  button100: string;
  button100Hover: string;
  button200: string;
  button200Hover: string;
  button300: string;
  button300Hover: string;
  button400: string;
  button400Hover: string;
  button500: string;
  button500Hover: string;
  button600: string;
  button600Hover: string;
  fillColor: string;
  backgroundColor: string;
  cautionColor: string;
}

export interface Theme {
  colors: colorThemeTypes;
}